pyproject.toml
[project]
name = "barqr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdf417>=0.8.1",
    "pdf417gen>=0.8.1",
    "pillow>=11.3.0",
    "pip>=25.1.1",
    "pydantic>=2.11.7",
    "python-barcode>=0.15.1",
    "qrcode>=8.2",
]
